added evaluation using ragas script
